Revert "Feat: Connect Socket for Game Start"
This reverts commit fb823cb3fb7ccd9690c2a23356ec03c3937a20b5.
