+++ src/components/Game/Keyword/index.tsx
@@ -1,4 +1,5 @@
 import {
+  Button,
   Center,
   Flex,
   Modal,
@@ -9,6 +10,7 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
+import data from "../../../data/category.json";
 
 interface Categories {
   category: string;
@@ -16,11 +18,13 @@ interface Categories {
 }
 [];
 
-const Keyword = ({ status }: any) => {
+const Keyword = ({ status, updateStatus }: any) => {
+  const categories = data.CategoryList;
   const [category, setCategory] = useState<Categories | null>(null);
   const [keyword, setKeyword] = useState("");
+  const users = data.users;
 
-  const { isOpen, onClose } = useDisclosure();
+  const { isOpen, onClose, onOpen } = useDisclosure();
 
   // 모달 자동 닫기
   useEffect(() => {
@@ -29,10 +33,39 @@ const Keyword = ({ status }: any) => {
     if (isOpen) {
       timer = setTimeout(() => {
         onClose();
-      }, 2500);
+      }, 3000);
     }
     return () => clearTimeout(timer);
   }, [isOpen, onClose]);
+
+  // 랜덤 숫자 계산
+  const getRandNum = (length: number): number => {
+    const randNum = Math.floor(Math.random() * length);
+
+    return randNum;
+  };
+
+  // 게임 시작
+  const handleStart = async () => {
+    await updateStatus("게임중");
+
+    const selectedCategory = categories[getRandNum(categories.length)];
+    const ranKeyword =
+      selectedCategory.keyword[getRandNum(selectedCategory.keyword.length)];
+    const ranLiar = users.name[getRandNum(users.name.length)];
+
+    setCategory(selectedCategory);
+    setKeyword(ranKeyword);
+
+    window.localStorage.setItem("category", selectedCategory.category);
+    window.localStorage.setItem("keyword", ranKeyword);
+    if (ranLiar === "연수") {
+      window.localStorage.setItem("liar", "true");
+    } else {
+      window.localStorage.setItem("liar", "false");
+    }
+    onOpen();
+  };
 
   // status 변화에 따라 localStorage에 저장
   useEffect(() => {
@@ -45,32 +78,56 @@ const Keyword = ({ status }: any) => {
     }
   }, [status]);
 
+  // 게임 종료
+  const hadleEnd = async () => {
+    await updateStatus("대기중");
+
+    window.localStorage.setItem("category", "");
+    window.localStorage.setItem("keyword", "");
+    window.localStorage.setItem("liar", "false");
+    setCategory(null);
+    setKeyword("");
+  };
+
   return (
-    <Modal isOpen={isOpen} onClose={onClose} closeOnEsc={true}>
-      <ModalOverlay />
-      <ModalContent>
-        <ModalCloseButton />
-        <ModalBody>
-          <Center fontWeight="bold" h="100%" fontSize="1.2rem" pt="20" pb="20">
-            <Flex
-              direction="column"
-              alignContent="center"
-              justifyContent="center"
+    <>
+      {status === "대기중" ? (
+        <Button onClick={handleStart}>게임시작</Button>
+      ) : (
+        <Button onClick={hadleEnd}>게임 종료</Button>
+      )}
+      <Modal isOpen={isOpen} onClose={onClose} closeOnEsc={true}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalCloseButton />
+          <ModalBody>
+            <Center
+              fontWeight="bold"
+              h="100%"
+              fontSize="1.2rem"
+              pt="20"
+              pb="20"
             >
-              <Center>주제는 {category?.category} 입니다.</Center>
-              {window.localStorage.getItem("liar") === "true" ? (
-                <>
-                  <Center>당신은 Liar 입니다.</Center>
-                  <Center>키워드를 추리하세요.</Center>
-                </>
-              ) : (
-                <Center>키워드는 {keyword} 입니다.</Center>
-              )}
-            </Flex>
-          </Center>
-        </ModalBody>
-      </ModalContent>
-    </Modal>
+              <Flex
+                direction="column"
+                alignContent="center"
+                justifyContent="center"
+              >
+                <Center>주제는 {category?.category} 입니다.</Center>
+                {window.localStorage.getItem("liar") === "true" ? (
+                  <>
+                    <Center>당신은 Liar 입니다.</Center>
+                    <Center>키워드를 추리하세요.</Center>
+                  </>
+                ) : (
+                  <Center>키워드는 {keyword} 입니다.</Center>
+                )}
+              </Flex>
+            </Center>
+          </ModalBody>
+        </ModalContent>
+      </Modal>
+    </>
   );
 };
